requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
SQLAlchemy==2.0.21
requests==2.31.0
schedule==1.2.0
supervisor==4.2.5
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
gunicorn==21.2.0